+++ src/main/java/com/taskadapter/redmineapi/bean/TimeEntry.java
@@ -1,6 +1,9 @@
 package com.taskadapter.redmineapi.bean;
 
+import java.util.Collection;
 import java.util.Date;
+import java.util.HashSet;
+import java.util.Set;
 
 public class TimeEntry implements Identifiable {
 
@@ -31,6 +34,7 @@ public class TimeEntry implements Identifiable {
     public final static Property<Date> SPENT_ON = new Property<>(Date.class, "spentOn");
     public final static Property<Date> CREATED_ON = new Property<>(Date.class, "createdOn");
     public final static Property<Date> UPDATED_ON = new Property<>(Date.class, "updatedOn");
+	public static final Property<Set<CustomField>> CUSTOM_FIELDS = new Property(Set.class, "custom_fields");
 
     /**
      * @param id database Id
@@ -38,6 +42,7 @@ public class TimeEntry implements Identifiable {
     TimeEntry(Integer id) {
         storage = new PropertyStorage();
         storage.set(DATABASE_ID, id);
+        storage.set(CUSTOM_FIELDS, new HashSet<>());
     }
 
     public Integer getId() {
@@ -140,11 +145,30 @@ public class TimeEntry implements Identifiable {
         storage.set(USER_NAME, userName);
     }
 
+    /**
+     * @return the custom field with given name or NULL if the field is not found
+     */
+    public CustomField getCustomField(String name) {
+    	return storage.get(CUSTOM_FIELDS).stream().filter(a -> a.getName().equals(name)).findFirst().orElse(null);
+    }
+
     @Override
     public String toString() {
         return "User \"" + getUserName() + "\" spent " + getHours()
                 + " hours on task " + getIssueId() + " (project \"" + getProjectName()
                 + "\") doing " + getActivityName();
+    }
+    
+    public Set<CustomField> getCustomFields() {
+    		return storage.get(CUSTOM_FIELDS);
+    }
+
+    public void clearCustomFields() {
+        storage.set(CUSTOM_FIELDS, new HashSet<>());
+    }
+
+    public void addCustomFields(Collection<CustomField> customFields) {
+    		storage.get(CUSTOM_FIELDS).addAll(customFields);
     }
 
     @Override